requirements.txt
click==7.1.1
Django==2.1
django-mysql>=3.2.0
django-graphql-jwt>=0.3.0
graphene-django>=2.0
sgqlc>=10.1
mysqlclient==1.3.13
python-dateutil>=2.8.0
grimoirelab_toolkit>=0.1.8
requests>=2.7.0
uri>=2.0.1
httpretty==0.9.7
jinja2==2.11.1
rq>=1.4.0
django-rq>=2.3.0
